refactor(i18n): drop ability to provide i18n init options
[sc-35331]
